Scoring fixed; oil generation in progress
Scoring now operates when drippy traverses an oil as opposed to when an
oil is generated. Oil generation however is tied with the scoring such
that the current oil generation factors cause the score to increase
exponentially. This section is mark with a comment stating “oil
generator”. This section involves a rather lengthy switch case.
Somewhere in how the case is applied, the score starts multiplying at
random.